requirements.txt
django-extensions
pyquery
requests
